refactor(config): move to cli package (#2405)
* chore: move config commands to cli package

* chore: move config tests to cli package

* chore: remove plugin-config import from cli package

* chore: move lib test to config directory and delete .keep file

* chore: remove config package

* fix: fix sinon import

* fix: remove .only
