+++ packages/cli/package.json
@@ -17,7 +17,6 @@
     "@heroku-cli/plugin-certs-v5": "^8.1.7",
     "@heroku-cli/plugin-ci": "^8.1.7",
     "@heroku-cli/plugin-ci-v5": "^8.1.8",
-    "@heroku-cli/plugin-config": "^8.1.7",
     "@heroku-cli/plugin-container-registry-v5": "^8.1.7",
     "@heroku-cli/plugin-git": "^8.1.8",
     "@heroku-cli/plugin-oauth-v5": "^8.1.9",
@@ -43,12 +42,14 @@
     "@oclif/plugin-which": "2.2.8",
     "date-fns": "^2.0.0-alpha.8",
     "debug": "4.1.1",
+    "edit-string": "^1.1.6",
     "execa": "5.1.1",
     "foreman": "^3.0.1",
     "fs-extra": "7.0.1",
     "http-call": "5.3.0",
     "netrc-parser": "3.1.6",
     "semver": "5.6.0",
+    "shell-quote": "^1.6.1",
     "uuid": "3.3.2"
   },
   "devDependencies": {
@@ -117,7 +118,6 @@
       "@heroku-cli/plugin-certs-v5",
       "@heroku-cli/plugin-ci-v5",
       "@heroku-cli/plugin-ci",
-      "@heroku-cli/plugin-config",
       "@heroku-cli/plugin-container-registry-v5",
       "@heroku-cli/plugin-git",
       "@heroku-cli/plugin-oauth-v5",
