+++ packages/cli/test/unit/commands/config/edit.unit.test.ts
@@ -1,8 +1,8 @@
 import {CliUx} from '@oclif/core'
-import sinon from 'sinon'
+import * as sinon from 'sinon'
 
 import {stringToConfig} from '../../../../src/commands/config/edit'
-import {Editor} from '../../../../src/util'
+import {Editor} from '../../../../src/lib/config/util'
 import {expect, test} from '@oclif/test'
 
 const cli = CliUx.ux
