--- packages/config/.gitattributes
@@ -1,3 +1,0 @@
-* text=auto
-*.js text eol=lf
-*.ts text eol=lf
