--- packages/config/.gitignore
@@ -1,9 +1,0 @@
-/.nyc_output
-*-debug.log
-*-error.log
-.oclif.manifest.json
-/dist
-/lib
-/package-lock.json
-/tmp
-node_modules
