--- packages/config/.nycrc
@@ -1,11 +1,0 @@
-{
-  "extension": [
-    ".ts"
-  ],
-  "include": [
-    "src/**/*.ts"
-  ],
-  "exclude": [
-    "**/*.d.ts"
-  ]
-}
