--- packages/config/bin/run
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-const oclif = require('@oclif/core')
-
-oclif.run().catch(require('@oclif/core/handle'))
