--- packages/config/package.json
@@ -1,68 +1,0 @@
-{
-  "name": "@heroku-cli/plugin-config",
-  "version": "8.1.7",
-  "author": "Jeff Dickey @jdxcode",
-  "bugs": "https://github.com/heroku/cli/issues",
-  "dependencies": {
-    "@heroku-cli/color": "^1.1.14",
-    "@heroku-cli/command": "^9.0.2",
-    "@oclif/core": "^1.26.2",
-    "edit-string": "^1.1.6",
-    "lodash": "^4.17.11",
-    "shell-quote": "^1.6.1"
-  },
-  "devDependencies": {
-    "@fancy-test/nock": "^0.1.1",
-    "@heroku-cli/schema": "^1.0.25",
-    "@oclif/test": "^2.2.20",
-    "@types/chai": "^4.1.7",
-    "@types/fs-extra": "^5.0.5",
-    "@types/lodash": "^4.14.123",
-    "@types/mocha": "^5.2.6",
-    "@types/nock": "^9.3.1",
-    "@types/supports-color": "^5.3.0",
-    "chai": "^4.2.0",
-    "globby": "^10.0.2",
-    "mocha": "^5.2.0",
-    "nock": "^10.0.6",
-    "nyc": "^15.1.0",
-    "oclif": "3.8.1",
-    "sinon": "^7.2.4",
-    "ts-node": "^10.9.1",
-    "tslib": "^1.9.3",
-    "typescript": "4.8.4"
-  },
-  "engines": {
-    "node": ">=14"
-  },
-  "files": [
-    "oclif.manifest.json",
-    "/lib"
-  ],
-  "homepage": "https://github.com/heroku/cli/tree/main/packages/config",
-  "keywords": [
-    "oclif-plugin"
-  ],
-  "license": "MIT",
-  "oclif": {
-    "additionalHelpFlags": [
-      "-h"
-    ],
-    "additionalVersionFlags": [
-      "-v"
-    ],
-    "bin": "heroku",
-    "commands": "./lib/commands",
-    "repositoryPrefix": "<%- repo %>/blob/v<%- version %>/packages/config/<%- commandPath %>"
-  },
-  "repository": "heroku/cli",
-  "scripts": {
-    "lint": "eslint . --ext .ts --config ../../.eslintrc --ignore-path ../../.eslintignore-lib",
-    "postpack": "rm -f oclif.manifest.json",
-    "prepack": "rm -rf lib && tsc && oclif manifest",
-    "pretest": "tsc -p test --noEmit",
-    "test": "nyc mocha --forbid-only \"test/**/*.unit.test.ts\"",
-    "posttest": "yarn lint",
-    "version": "oclif readme && git add README.md"
-  }
-}
