--- packages/config/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "forceConsistentCasingInFileNames": true,
-    "importHelpers": true,
-    "esModuleInterop": true,
-    "module": "commonjs",
-    "outDir": "./lib",
-    "pretty": true,
-    "rootDirs": [
-      "./src"
-    ],
-    "skipLibCheck": true,
-    "strict": true,
-    "target": "es2017"
-  },
-  "include": [
-    "./src/**/*"
-  ]
-}
